backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.48.2",
    "fastapi[standard]>=0.133.1",
    "loguru>=0.7.3",
    "openai>=2.24.0",
    "pydantic-settings>=2.13.1",
]
